Project import generated by Copybara.
GitOrigin-RevId: 30f49efad73c35cb9adad172f6854f1709527dba

+++ native/shufflenet_interface/prebuilt/megengine/include/megbrain/serialization/opr_registry.h
@@ -76,7 +76,7 @@ namespace serialization {
 
 #if MGB_ENABLE_DEBUG_UTIL
         //! dump registered oprs
-        static std::vector<std::pair<uint64_t, std::string>> dump_registries();
+        static std::vector<std::pair<size_t, std::string>> dump_registries();
 #endif
     };
